Display first emblem as page & disable init query on it
